refactor: move color to own package
