+++ internal/pkg/flink/components/interactive_input.go
@@ -3,9 +3,9 @@ package components
 import (
 	"strings"
 
-	"github.com/fatih/color"
+	fColor "github.com/fatih/color"
 
-	"github.com/confluentinc/cli/internal/pkg/flink/config"
+	"github.com/confluentinc/cli/internal/pkg/color"
 	"github.com/confluentinc/cli/internal/pkg/output"
 )
 
@@ -25,7 +25,7 @@ func PrintOptionState(prefix string, isEnabled bool, maxCol int) {
 		stateMsg = "enabled"
 	}
 
-	output.Printf("\n" + prefix + color.CyanString(stateMsg))
+	output.Printf("\n" + prefix + fColor.CyanString(stateMsg))
 
 	// This prints to the console the exact amount of empty characters to fill the line might have autocompletions before
 	output.Println(strings.Repeat(" ", maxCol-len(prefix+stateMsg)))
@@ -37,6 +37,6 @@ func PrintWelcomeHeader() {
 	output.Printf("To exit, press Ctrl-Q or type \"exit\". \n\n")
 
 	// Print shortcuts
-	c := color.New(config.AccentColor)
+	c := fColor.New(color.AccentColor)
 	output.Printf("[Ctrl-Q] %s [Ctrl-S] %s \n", c.Sprint("Quit"), c.Sprint("Toggle Smart Completion"))
 }
