+++ internal/pkg/flink/internal/controller/input_controller_utils.go
@@ -3,12 +3,12 @@ package controller
 import (
 	"os"
 
-	"github.com/fatih/color"
+	fColor "github.com/fatih/color"
 	"golang.org/x/term"
 
 	"github.com/confluentinc/go-prompt"
 
-	"github.com/confluentinc/cli/internal/pkg/flink/config"
+	"github.com/confluentinc/cli/internal/pkg/color"
 	"github.com/confluentinc/cli/internal/pkg/log"
 	"github.com/confluentinc/cli/internal/pkg/output"
 )
@@ -45,26 +45,26 @@ func restoreStdin(state *term.State) {
 }
 
 func outputErr(s string) {
-	c := color.New(config.ErrorColor)
+	c := fColor.New(color.ErrorColor)
 	output.Println(c.Sprintf(s))
 }
 
 func outputErrf(s string, args ...any) {
-	c := color.New(config.ErrorColor)
+	c := fColor.New(color.ErrorColor)
 	output.Printf(c.Sprint(s), args...)
 }
 
 func outputInfo(s string) {
-	c := color.New(config.InfoColor)
+	c := fColor.New(color.InfoColor)
 	output.Println(c.Sprint(s))
 }
 
 func outputInfof(s string, args ...any) {
-	c := color.New(config.InfoColor)
+	c := fColor.New(color.InfoColor)
 	output.Printf(c.Sprint(s), args...)
 }
 
 func outputWarn(s string) {
-	c := color.New(config.WarnColor)
+	c := fColor.New(color.WarnColor)
 	output.Println(c.Sprint(s))
 }
